refactor(cache): restructure cache layer
